# stream: reuse buffers