--- a/processor/stream/stream_processor.go
+++ b/processor/stream/stream_processor.go
@@ -22,6 +22,7 @@ import (
 	"context"
 	"errors"
 	"io"
+	"sync"
 	"time"
 
 	"github.com/santhosh-tekuri/jsonschema"
@@ -42,6 +43,7 @@ import (
 
 var (
 	ErrUnrecognizedObject = errors.New("did not recognize object type")
+	bufferPool            sync.Pool
 )
 
 type StreamReader interface {
@@ -234,7 +236,17 @@ func (s *StreamProcessor) readBatch(ctx context.Context, rl *rate.Limiter, batch
 func (s *StreamProcessor) HandleStream(ctx context.Context, rl *rate.Limiter, meta map[string]interface{}, reader io.Reader, report publish.Reporter) *Result {
 	res := &Result{}
 
-	buf := bufio.NewReaderSize(reader, s.MaxEventSize)
+	buf, ok := bufferPool.Get().(*bufio.Reader)
+	if !ok {
+		buf = bufio.NewReaderSize(reader, s.MaxEventSize)
+	} else {
+		buf.Reset(reader)
+	}
+	defer func() {
+		buf.Reset(nil)
+		bufferPool.Put(buf)
+	}()
+
 	lineReader := decoder.NewLineReader(buf, s.MaxEventSize)
 	ndReader := decoder.NewNDJSONStreamReader(lineReader)