Following up on roncohen/apm-server#4 here are fresh benchmarks:

**ns/op:**
```
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      244409        170019        -30.44%
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    264594        201918        -23.69%
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      132161        71012         -46.27%
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    163831        87221         -46.76%
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       24840950      21281166      -14.33%
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     30415723      26425746      -13.12%
```

**MB/s:**
```
benchmark                                                                 old MB/s     new MB/s     speedup
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      18.01        25.89        1.44x
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    16.64        21.80        1.31x
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      7.62         14.18        1.86x
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    6.15         11.55        1.88x
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       16.05        18.73        1.17x
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     13.10        15.08        1.15x
```

**bytes/op:**
```
benchmark                                                                 old bytes     new bytes     delta
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      361925        55522         -84.66%
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    362460        55924         -84.57%
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      334528        25434         -92.40%
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    335060        26344         -92.14%
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       7745830       7441515       -3.93%
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     7847380       7548856       -3.80%
```



<details>

```
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      244409        170019        -30.44%
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    264594        201918        -23.69%
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      132161        71012         -46.27%
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    163831        87221         -46.76%
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       24840950      21281166      -14.33%
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     30415723      26425746      -13.12%
BenchmarkStreamProcessor/invalid-event.ndjson/NoRateLimit-4               130544        45298         -65.30%
BenchmarkStreamProcessor/invalid-event.ndjson/WithRateLimit-4             152158        62069         -59.21%
BenchmarkStreamProcessor/invalid-json-event.ndjson/NoRateLimit-4          90355         36047         -60.11%
BenchmarkStreamProcessor/invalid-json-event.ndjson/WithRateLimit-4        142713        69408         -51.37%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/NoRateLimit-4       41283         5564          -86.52%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/WithRateLimit-4     37814         5818          -84.61%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/NoRateLimit-4          37682         5826          -84.54%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/WithRateLimit-4        37698         5633          -85.06%
BenchmarkStreamProcessor/invalid-metadata.ndjson/NoRateLimit-4            44358         8942          -79.84%
BenchmarkStreamProcessor/invalid-metadata.ndjson/WithRateLimit-4          42398         8957          -78.87%
BenchmarkStreamProcessor/metadata-null-values.ndjson/NoRateLimit-4        98614         31619         -67.94%
BenchmarkStreamProcessor/metadata-null-values.ndjson/WithRateLimit-4      127285        49150         -61.39%
BenchmarkStreamProcessor/metricsets.ndjson/NoRateLimit-4                  116513        61841         -46.92%
BenchmarkStreamProcessor/metricsets.ndjson/WithRateLimit-4                164252        76707         -53.30%
BenchmarkStreamProcessor/minimal-service.ndjson/NoRateLimit-4             93606         36447         -61.06%
BenchmarkStreamProcessor/minimal-service.ndjson/WithRateLimit-4           123373        53315         -56.79%
BenchmarkStreamProcessor/only-metadata.ndjson/NoRateLimit-4               75254         30661         -59.26%
BenchmarkStreamProcessor/only-metadata.ndjson/WithRateLimit-4             104885        46102         -56.05%
BenchmarkStreamProcessor/optional-timestamps.ndjson/NoRateLimit-4         121416        61813         -49.09%
BenchmarkStreamProcessor/optional-timestamps.ndjson/WithRateLimit-4       148060        78969         -46.66%
BenchmarkStreamProcessor/ratelimit.ndjson/NoRateLimit-4                   318699        250314        -21.46%
BenchmarkStreamProcessor/ratelimit.ndjson/WithRateLimit-4                 363269        288604        -20.55%
BenchmarkStreamProcessor/spans.ndjson/NoRateLimit-4                       205056        135608        -33.87%
BenchmarkStreamProcessor/spans.ndjson/WithRateLimit-4                     222463        151135        -32.06%
BenchmarkStreamProcessor/transactions.ndjson/NoRateLimit-4                185118        128713        -30.47%
BenchmarkStreamProcessor/transactions.ndjson/WithRateLimit-4              200850        148688        -25.97%
BenchmarkStreamProcessor/unrecognized-event.ndjson/NoRateLimit-4          67515         23651         -64.97%
BenchmarkStreamProcessor/unrecognized-event.ndjson/WithRateLimit-4        95973         40807         -57.48%

benchmark                                                                 old MB/s     new MB/s     speedup
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      18.01        25.89        1.44x
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    16.64        21.80        1.31x
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      7.62         14.18        1.86x
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    6.15         11.55        1.88x
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       16.05        18.73        1.17x
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     13.10        15.08        1.15x
BenchmarkStreamProcessor/invalid-event.ndjson/NoRateLimit-4               5.87         16.91        2.88x
BenchmarkStreamProcessor/invalid-event.ndjson/WithRateLimit-4             5.03         12.34        2.45x
BenchmarkStreamProcessor/invalid-json-event.ndjson/NoRateLimit-4          6.50         16.28        2.50x
BenchmarkStreamProcessor/invalid-json-event.ndjson/WithRateLimit-4        4.11         8.46         2.06x
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/NoRateLimit-4       11.09        82.31        7.42x
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/WithRateLimit-4     12.11        78.72        6.50x
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/NoRateLimit-4          11.89        76.89        6.47x
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/WithRateLimit-4        11.88        79.52        6.69x
BenchmarkStreamProcessor/invalid-metadata.ndjson/NoRateLimit-4            10.30        51.10        4.96x
BenchmarkStreamProcessor/invalid-metadata.ndjson/WithRateLimit-4          10.78        51.02        4.73x
BenchmarkStreamProcessor/metadata-null-values.ndjson/NoRateLimit-4        5.04         15.72        3.12x
BenchmarkStreamProcessor/metadata-null-values.ndjson/WithRateLimit-4      3.90         10.11        2.59x
BenchmarkStreamProcessor/metricsets.ndjson/NoRateLimit-4                  7.60         14.33        1.89x
BenchmarkStreamProcessor/metricsets.ndjson/WithRateLimit-4                5.39         11.55        2.14x
BenchmarkStreamProcessor/minimal-service.ndjson/NoRateLimit-4             4.54         11.66        2.57x
BenchmarkStreamProcessor/minimal-service.ndjson/WithRateLimit-4           3.44         7.97         2.32x
BenchmarkStreamProcessor/only-metadata.ndjson/NoRateLimit-4               7.43         18.23        2.45x
BenchmarkStreamProcessor/only-metadata.ndjson/WithRateLimit-4             5.33         12.13        2.28x
BenchmarkStreamProcessor/optional-timestamps.ndjson/NoRateLimit-4         8.46         16.61        1.96x
BenchmarkStreamProcessor/optional-timestamps.ndjson/WithRateLimit-4       6.94         13.01        1.87x
BenchmarkStreamProcessor/ratelimit.ndjson/NoRateLimit-4                   13.22        16.83        1.27x
BenchmarkStreamProcessor/ratelimit.ndjson/WithRateLimit-4                 11.59        14.59        1.26x
BenchmarkStreamProcessor/spans.ndjson/NoRateLimit-4                       12.95        19.59        1.51x
BenchmarkStreamProcessor/spans.ndjson/WithRateLimit-4                     11.94        17.57        1.47x
BenchmarkStreamProcessor/transactions.ndjson/NoRateLimit-4                15.27        21.96        1.44x
BenchmarkStreamProcessor/transactions.ndjson/WithRateLimit-4              14.07        19.01        1.35x
BenchmarkStreamProcessor/unrecognized-event.ndjson/NoRateLimit-4          5.79         16.53        2.85x
BenchmarkStreamProcessor/unrecognized-event.ndjson/WithRateLimit-4        4.07         9.58         2.35x

benchmark                                                                 old allocs     new allocs     delta
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      947            943            -0.42%
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    955            951            -0.42%
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      377            374            -0.80%
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    385            382            -0.78%
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       134572         134573         +0.00%
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     136146         136144         -0.00%
BenchmarkStreamProcessor/invalid-event.ndjson/NoRateLimit-4               268            263            -1.87%
BenchmarkStreamProcessor/invalid-event.ndjson/WithRateLimit-4             276            271            -1.81%
BenchmarkStreamProcessor/invalid-json-event.ndjson/NoRateLimit-4          199            196            -1.51%
BenchmarkStreamProcessor/invalid-json-event.ndjson/WithRateLimit-4        215            212            -1.40%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/NoRateLimit-4       29             27             -6.90%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/WithRateLimit-4     29             27             -6.90%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/NoRateLimit-4          24             22             -8.33%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/WithRateLimit-4        24             22             -8.33%
BenchmarkStreamProcessor/invalid-metadata.ndjson/NoRateLimit-4            53             49             -7.55%
BenchmarkStreamProcessor/invalid-metadata.ndjson/WithRateLimit-4          53             49             -7.55%
BenchmarkStreamProcessor/metadata-null-values.ndjson/NoRateLimit-4        178            175            -1.69%
BenchmarkStreamProcessor/metadata-null-values.ndjson/WithRateLimit-4      186            183            -1.61%
BenchmarkStreamProcessor/metricsets.ndjson/NoRateLimit-4                  296            293            -1.01%
BenchmarkStreamProcessor/metricsets.ndjson/WithRateLimit-4                304            301            -0.99%
BenchmarkStreamProcessor/minimal-service.ndjson/NoRateLimit-4             201            198            -1.49%
BenchmarkStreamProcessor/minimal-service.ndjson/WithRateLimit-4           209            206            -1.44%
BenchmarkStreamProcessor/only-metadata.ndjson/NoRateLimit-4               212            209            -1.42%
BenchmarkStreamProcessor/only-metadata.ndjson/WithRateLimit-4             220            217            -1.36%
BenchmarkStreamProcessor/optional-timestamps.ndjson/NoRateLimit-4         388            385            -0.77%
BenchmarkStreamProcessor/optional-timestamps.ndjson/WithRateLimit-4       396            393            -0.76%
BenchmarkStreamProcessor/ratelimit.ndjson/NoRateLimit-4                   1448           1444           -0.28%
BenchmarkStreamProcessor/ratelimit.ndjson/WithRateLimit-4                 1465           1461           -0.27%
BenchmarkStreamProcessor/spans.ndjson/NoRateLimit-4                       767            764            -0.39%
BenchmarkStreamProcessor/spans.ndjson/WithRateLimit-4                     776            772            -0.52%
BenchmarkStreamProcessor/transactions.ndjson/NoRateLimit-4                657            654            -0.46%
BenchmarkStreamProcessor/transactions.ndjson/WithRateLimit-4              666            662            -0.60%
BenchmarkStreamProcessor/unrecognized-event.ndjson/NoRateLimit-4          129            126            -2.33%
BenchmarkStreamProcessor/unrecognized-event.ndjson/WithRateLimit-4        137            134            -2.19%

benchmark                                                                 old bytes     new bytes     delta
BenchmarkStreamProcessor/errors.ndjson/NoRateLimit-4                      361925        55522         -84.66%
BenchmarkStreamProcessor/errors.ndjson/WithRateLimit-4                    362460        55924         -84.57%
BenchmarkStreamProcessor/events.ndjson/NoRateLimit-4                      334528        25434         -92.40%
BenchmarkStreamProcessor/events.ndjson/WithRateLimit-4                    335060        26344         -92.14%
BenchmarkStreamProcessor/heavy.ndjson/NoRateLimit-4                       7745830       7441515       -3.93%
BenchmarkStreamProcessor/heavy.ndjson/WithRateLimit-4                     7847380       7548856       -3.80%
BenchmarkStreamProcessor/invalid-event.ndjson/NoRateLimit-4               327064        17109         -94.77%
BenchmarkStreamProcessor/invalid-event.ndjson/WithRateLimit-4             327632        17532         -94.65%
BenchmarkStreamProcessor/invalid-json-event.ndjson/NoRateLimit-4          324103        13880         -95.72%
BenchmarkStreamProcessor/invalid-json-event.ndjson/WithRateLimit-4        325175        15138         -95.34%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/NoRateLimit-4       313344        2181          -99.30%
BenchmarkStreamProcessor/invalid-json-metadata.ndjson/WithRateLimit-4     313344        2176          -99.31%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/NoRateLimit-4          313184        2036          -99.35%
BenchmarkStreamProcessor/invalid-metadata-2.ndjson/WithRateLimit-4        313184        2030          -99.35%
BenchmarkStreamProcessor/invalid-metadata.ndjson/NoRateLimit-4            314609        3393          -98.92%
BenchmarkStreamProcessor/invalid-metadata.ndjson/WithRateLimit-4          314610        3396          -98.92%
BenchmarkStreamProcessor/metadata-null-values.ndjson/NoRateLimit-4        323007        12671         -96.08%
BenchmarkStreamProcessor/metadata-null-values.ndjson/WithRateLimit-4      323550        13076         -95.96%
BenchmarkStreamProcessor/metricsets.ndjson/NoRateLimit-4                  330392        20639         -93.75%
BenchmarkStreamProcessor/metricsets.ndjson/WithRateLimit-4                330918        21374         -93.54%
BenchmarkStreamProcessor/minimal-service.ndjson/NoRateLimit-4             325358        15337         -95.29%
BenchmarkStreamProcessor/minimal-service.ndjson/WithRateLimit-4           325898        16003         -95.09%
BenchmarkStreamProcessor/only-metadata.ndjson/NoRateLimit-4               325346        15394         -95.27%
BenchmarkStreamProcessor/only-metadata.ndjson/WithRateLimit-4             325883        15759         -95.16%
BenchmarkStreamProcessor/optional-timestamps.ndjson/NoRateLimit-4         334812        26033         -92.22%
BenchmarkStreamProcessor/optional-timestamps.ndjson/WithRateLimit-4       335340        26708         -92.04%
BenchmarkStreamProcessor/ratelimit.ndjson/NoRateLimit-4                   393138        87139         -77.84%
BenchmarkStreamProcessor/ratelimit.ndjson/WithRateLimit-4                 394236        88588         -77.53%
BenchmarkStreamProcessor/spans.ndjson/NoRateLimit-4                       353665        46469         -86.86%
BenchmarkStreamProcessor/spans.ndjson/WithRateLimit-4                     354192        47149         -86.69%
BenchmarkStreamProcessor/transactions.ndjson/NoRateLimit-4                349479        42371         -87.88%
BenchmarkStreamProcessor/transactions.ndjson/WithRateLimit-4              350016        42641         -87.82%
BenchmarkStreamProcessor/unrecognized-event.ndjson/NoRateLimit-4          321313        10784         -96.64%
BenchmarkStreamProcessor/unrecognized-event.ndjson/WithRateLimit-4        321845        11405         -96.46%
```
</details>